refactor: split monolithic index.ts into modular lib/ structure
Restructure the extension from a single 690-line file into 11
well-organized modules. No logic changes — purely structural.

- lib/types.ts          — all type/interface definitions
- lib/constants.ts      — provider IDs, ports, TTLs
- lib/credentials.ts    — OAuth credential encode/decode
- lib/url-helpers.ts    — buildBaseUrl, authHeaders
- lib/discovery.ts      — UDP beacon + HTTP port scanning
- lib/http.ts           — /health and /models API calls
- lib/models.ts         — model mapping (Lemonade → Pi shape)
- lib/provider.ts       — provider (re-)registration
- lib/oauth.ts          — /login OAuth flow
- lib/admin.ts          — /lemonade admin command
- extensions/index.ts   — thin entry point (70 lines)

No behavior changes. All imports resolve correctly. No circular
dependencies.
